+++ packages/block-library/src/icon/edit.js
@@ -1,4 +1,9 @@
 /**
+ * External dependencies
+ */
+import clsx from 'clsx';
+
+/**
  * WordPress dependencies
  */
 import { __ } from '@wordpress/i18n';
@@ -16,6 +21,10 @@ import {
 	InspectorControls,
 	useBlockProps,
 	useBlockEditingMode,
+	__experimentalUseColorProps as useColorProps,
+	__experimentalUseBorderProps as useBorderProps,
+	__experimentalGetSpacingClassesAndStyles as useSpacingProps,
+	getDimensionsClassesAndStyles as useDimensionsProps,
 	store as blockEditorStore,
 } from '@wordpress/block-editor';
 import { useDispatch, useSelect } from '@wordpress/data';
@@ -40,11 +49,16 @@ export function Edit( { attributes, setAttributes } ) {
 
 	const isContentOnlyMode = useBlockEditingMode() === 'contentOnly';
 
+	const colorProps = useColorProps( attributes );
+	const spacingProps = useSpacingProps( attributes );
+	const borderProps = useBorderProps( attributes );
+	const dimensionsProps = useDimensionsProps( attributes );
+
 	const allIcons = useSelect( ( select ) => {
 		return unlock( select( coreDataStore ) ).getIcons();
 	}, [] );
 
-	// Is the width value is 0, reset it to the default value.
+	// Is the width value is 0, reset it to the minimum value.
 	useEffect( () => {
 		if (
 			! style?.dimensions?.width ||
@@ -161,13 +175,36 @@ export function Edit( { attributes, setAttributes } ) {
 			{ inspectorControls }
 			<div { ...useBlockProps() }>
 				{ icon ? (
-					<HtmlRenderer html={ iconToDisplay } />
+					<HtmlRenderer
+						html={ iconToDisplay }
+						wrapperProps={ {
+							className: clsx(
+								colorProps.className,
+								borderProps.className,
+								spacingProps.className,
+								dimensionsProps.className
+							),
+							style: {
+								...colorProps.style,
+								...borderProps.style,
+								...spacingProps.style,
+								...dimensionsProps.style,
+							},
+						} }
+					/>
 				) : (
 					<Placeholder
 						withIllustration
+						className={ clsx(
+							borderProps.className,
+							spacingProps.className,
+							dimensionsProps.className
+						) }
 						style={ {
-							height: attributes?.style?.dimensions?.width,
-							width: attributes?.style?.dimensions?.width,
+							...borderProps.style,
+							...spacingProps.style,
+							...dimensionsProps.style,
+							aspectRatio: '1',
 						} }
 					/>
 				) }